Add '*' & '/' instead of 'x' & '÷' to displayValue
Previously when an operator button was pressed depending on its id, the
corresponding symbol would be stored in the operator variable while its
text content would be stored in the displayValue variable.

This caused problems as the symbols for times and divide are different
than the text content of their buttons, this meant that the regular
expression [*-+/] could not match and remove '*' '/' in the displayValue
string when split. This would then later cause Nan to be returned when
trying to convert '*' and '/' character to numbers.